Boilerplate for new trunk system
